MindSpore-Retrieval-augmented/requirements.txt
Package            Version
------------------ ------------
aiohttp            3.9.0
aiosignal          1.3.1
asttokens          2.4.1
astunparse         1.6.3
async-timeout      4.0.3
attrs              23.1.0
certifi            2023.11.17
charset-normalizer 3.3.2
datasets           2.15.0
dill               0.3.7
faiss-cpu          1.7.4
filelock           3.13.1
frozenlist         1.4.0
fsspec             2023.10.0
huggingface-hub    0.19.4
idna               3.4
mindnlp            0.2.0
mindspore          2.2.0
multidict          6.0.4
multiprocess       0.70.15
numpy              1.24.4
packaging          23.2
pandas             2.0.3
Pillow             9.5.0
pip                23.3
protobuf           4.25.1
psutil             5.9.6
pyarrow            14.0.1
pyarrow-hotfix     0.5
python-dateutil    2.8.2
pytz               2023.3.post1
PyYAML             6.0.1
regex              2023.10.3
requests           2.31.0
scipy              1.10.1
sentencepiece      0.1.99
setuptools         68.0.0
six                1.16.0
tokenizers         0.15.0
tqdm               4.66.1
typing_extensions  4.8.0
tzdata             2023.3
urllib3            2.1.0
wheel              0.41.2
xxhash             3.4.1
yarl               1.9.2
